Extract LoadInspectorBase class
